+++ packages/react-select-async-paginate/src/types.ts
@@ -109,6 +109,10 @@ export type UseAsyncPaginateParams<
 	 * Clear all cached options on search change
 	 */
 	clearCacheOnSearchChange?: boolean;
+	/**
+	 * Clear all cached options on menu close
+	 */
+	clearCacheOnMenuClose?: boolean;
 	loadOptions: LoadOptions<OptionType, Group, Additional>;
 	options?: OptionsOrGroups<OptionType, Group>;
 	defaultOptions?: boolean | OptionsOrGroups<OptionType, Group>;